Main gemaakt met wat tests, ook P4 gecorrigeerd

+++ src/main/java/Model/Product.java
@@ -7,10 +7,10 @@ public class Product {
     private int productNummer;
     private String naam;
     private String beschrijving;
-    private int prijs;
+    private double prijs;
     public List<OVChipkaart> ovChipList;
 
-    public Product(int pNum, String naam, String bschr, int prijs) {
+    public Product(int pNum, String naam, String bschr, double prijs) {
         this.productNummer = pNum;
         this.naam = naam;
         this.beschrijving = bschr;
@@ -26,7 +26,7 @@ public class Product {
     public void setBeschrijving(String beschrijving) {
         this.beschrijving = beschrijving;
     }
-    public void setPrijs(int prijs) {
+    public void setPrijs(double prijs) {
         this.prijs = prijs;
     }
 
@@ -39,7 +39,7 @@ public class Product {
     public String getBeschrijving() {
         return beschrijving;
     }
-    public int getPrijs() {
+    public double getPrijs() {
         return prijs;
     }
     public List<OVChipkaart> getOvChipList() {
